Re-organize into module structure
