+++ src/event.rs
@@ -1,7 +1,9 @@
 use std::time::Instant;
 use std::path::PathBuf;
 
-use {DeviceId, LogicalPosition, LogicalSize, WindowId};
+use dpi::{LogicalPosition, LogicalSize};
+use window::WindowId;
+use platform_impl;
 
 /// Describes a generic event.
 #[derive(Clone, Debug)]
@@ -142,6 +144,14 @@ pub enum WindowEvent {
     /// For more information about DPI in general, see the [`dpi`](dpi/index.html) module.
     HiDpiFactorChanged(f64),
 }
+
+/// Identifier of an input device.
+///
+/// Whenever you receive an event arising from a particular input device, this event contains a `DeviceId` which
+/// identifies its origin. Note that devices may be virtual (representing an on-screen cursor and keyboard focus) or
+/// physical. Virtual devices typically aggregate inputs from multiple physical devices.
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
+pub struct DeviceId(pub(crate) platform_impl::DeviceId);
 
 /// Represents raw hardware events that are not associated with any particular window.
 ///
