+++ tests/send_objects.rs
@@ -5,19 +5,19 @@ fn needs_send<T:Send>() {}
 #[test]
 fn events_loop_proxy_send() {
     // ensures that `winit::EventLoopProxy` implements `Send`
-    needs_send::<winit::EventLoopProxy<()>>();
+    needs_send::<winit::event_loop::EventLoopProxy<()>>();
 }
 
 #[test]
 fn window_send() {
     // ensures that `winit::Window` implements `Send`
-    needs_send::<winit::Window>();
+    needs_send::<winit::window::Window>();
 }
 
 #[test]
 fn ids_send() {
     // ensures that the various `..Id` types implement `Send`
-    needs_send::<winit::WindowId>();
-    needs_send::<winit::DeviceId>();
-    needs_send::<winit::MonitorId>();
+    needs_send::<winit::window::WindowId>();
+    needs_send::<winit::event::DeviceId>();
+    needs_send::<winit::monitor::MonitorId>();
 }
