+++ tests/sync_object.rs
@@ -5,5 +5,5 @@ fn needs_sync<T:Sync>() {}
 #[test]
 fn window_sync() {
     // ensures that `winit::Window` implements `Sync`
-    needs_sync::<winit::Window>();
+    needs_sync::<winit::window::Window>();
 }
